pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillforge"
version = "0.1.0"
description = "Build, test, and publish AI agent skills across every platform."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Devashish", email = "jadhavom263@gmail.com" }]
keywords = ["ai", "agents", "skills", "clawhub", "claude", "mcp", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "rich>=13.7",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "litellm>=1.40",
    "httpx>=0.27",
    "pynacl>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "typer[all]",
]

[project.urls]
Homepage   = "https://github.com/ddevilz/skillforge"
Repository = "https://github.com/ddevilz/skillforge"
Issues     = "https://github.com/ddevilz/skillforge/issues"
Docs       = "https://skillforge.dev"

[project.scripts]
skillforge = "skillforge.cli:entry"

[tool.hatch.build.targets.wheel]
packages = ["skillforge"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "skillforge/templates"

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "typer[all]",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths   = ["tests"]
addopts     = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
